requirements.txt
streamlit
Pillow
python-barcode
openpyxl
reportlab
pypdf
gspread
google-auth
plotly
